fix(lint): clear clippy errors and warnings so CI passes under -D warnings
CI runs `cargo clippy --workspace --exclude volterra-py -- -D warnings`,
which was failing on 3 errors and 24 warnings.

Errors (all false positives on arbitrary test sentinels that merely
resembled PI/E): change 2.71828 to 2.65 in a handcrafted rk4 test field,
and the 3.14 "untouched cell" marker to 42.0 in test_bc.

Warnings:
- too_many_arguments: targeted #[allow] on 7 FD-kernel / physics-driver
  functions (splitting their signatures would breach the no-API-change
  goal; they legitimately take many physical and grid parameters)
- large_const_arrays: golden ORACLE_* arrays changed const -> static
- redundant i64->i64 casts on params.max_p_iters dropped (+ doc fixup)
- manual RangeInclusive::contains, map_or -> is_none_or, iter().copied()
  .collect() -> to_vec(), let-binding return, unused binding, needless
  range loop, and an intentional const-assertion guard: all cleared

No numerical scheme or public API change; golden tests stay bit-identical.

+++ volterra-cgpo/src/step.rs
@@ -208,7 +208,7 @@ pub fn update_step_inner(
         params,
         bnd,
         target_rel_change,
-        params.max_p_iters as i64,
+        params.max_p_iters,
     );
 
     // 5. Q update
